shifter tests moved and passing
